refactor into 1 inscription method
