--- src/inscriptions/textInscription.ts
@@ -1,22 +1,0 @@
-import type { Json } from 'jsontokens';
-import { createUnsecuredToken } from 'jsontokens';
-
-import { getProviderOrThrow } from '../provider';
-import { validateInscriptionPayload } from './common';
-import { CreateTextInscriptionOptions } from './types';
-
-export const createTextInscription = async (options: CreateTextInscriptionOptions) => {
-  const { getProvider } = options;
-  const provider = await getProviderOrThrow(getProvider);
-
-  validateInscriptionPayload(options.payload);
-
-  try {
-    const request = createUnsecuredToken(options.payload as unknown as Json);
-    const response = await provider.createTextInscription(request);
-    options.onFinish?.(response);
-  } catch (error) {
-    console.error('[Connect] Error during address request', error);
-    options.onCancel?.();
-  }
-};
